fix: nested types generation
- Generated file name include containing type name.
- Reference nested type correctly for JSON converter.
- Avoid generating to file-scoped enum.
- Generate documentation file for NuGet package.
