+++ gen/EnumUtilities.Generators.Roslyn5_0_0/CodeWriters/EnumJsonConverterWriter.cs
@@ -1,4 +1,5 @@
-﻿using Raiqub.Generators.EnumUtilities.Common;
+﻿using Microsoft.CodeAnalysis.CSharp;
+using Raiqub.Generators.EnumUtilities.Common;
 using Raiqub.Generators.EnumUtilities.Models;
 using Raiqub.Generators.InterpolationCodeWriter;
 
@@ -41,12 +42,12 @@ public sealed class EnumJsonConverterWriter : ICodeWriter<EnumToGenerate>
 
         writer.WriteLine(
             $$"""
-              {{(
-                  model.IsPublic ? "public" : "internal"
-              )}} sealed partial class {{model.Name}}JsonConverter : JsonConverter<{{model.Name}}>
-              {
-                  private const int MaxCharStack = 256;
-              """
+            {{SyntaxFacts.GetText(
+                model.AccessModifier
+            )}} sealed partial class {{model.Name}}JsonConverter : JsonConverter<{{model.RefName}}>
+            {
+                private const int MaxCharStack = 256;
+            """
         );
         writer.PushIndent();
 
@@ -211,7 +212,7 @@ public sealed class EnumJsonConverterWriter : ICodeWriter<EnumToGenerate>
 
     private static void WriteReadFromNumber(SourceTextWriter writer, EnumToGenerate model)
     {
-        var underlyingTypeName = CSharpExtensions.GetTypeNameFromCSharpKeyword(model.UnderlyingType);
+        var underlyingTypeName = CSharpUtils.GetTypeNameFromCSharpKeyword(model.UnderlyingType);
         writer.WriteLine(
             $$"""
             private {{model.UnderlyingType}} ReadFromNumber(ref Utf8JsonReader reader)
