+++ tests/EnumUtilities.Generators.IntegrationTests/EnumUtilities.Generators.IntegrationTests.csproj
@@ -33,7 +33,7 @@
 
   <ItemGroup Condition="'$(TargetFramework)' == 'net6.0'">
     <ProjectReference
-      Include="..\..\gen\EnumUtilities.Generators.Roslyn4_8_0\EnumUtilities.Generators.Roslyn4_8_0.csproj"
+      Include="..\..\gen\EnumUtilities.Generators.Roslyn4_3_1\EnumUtilities.Generators.Roslyn4_3_1.csproj"
       OutputItemType="Analyzer"
       ReferenceOutputAssembly="false"
     />
